Implemented method to pass third test
To replace multiples of three with "Fizz", multiples of five with "Buzz", and multiples of both three and five with "FizzBuzz", I used if statements which check if the number of the current iteration is divisible by three but not five, five but not three, and fifteen, respectively. If the condition is met, the appropriate replacement is concatenated to the string.