Implemented method to pass first test
To output whether or not a given year is a leap year, I used an if statement to check if the given year is divisible by four. If so, the code returns a string saying that the given year is a leap year. Otherwise, the code returns a string saying that the given year is not a leap year.